PPS-63: Refactor styled-elements
