+++ src/components/styled-elements/components/StyledOverlay.js
