+++ src/components/styled-elements/components/StyledPanel.js
